+++ packages/cap-config-writer/src/cap-config/package-json.ts
@@ -2,8 +2,8 @@ import { coerce, gte, satisfies } from 'semver';
 import { getCapCustomPaths } from '@sap-ux/project-access';
 import type { Package } from '@sap-ux/project-access';
 
-const minCdsVersion = '6.8.2';
-const minCdsPluginUi5Version = '0.6.13';
+export const minCdsVersion = '6.8.2';
+const minCdsPluginUi5Version = '0.9.3';
 
 /**
  * Ensure a minimum version of @sap/cds in dependencies.