+++ packages/fiori-annotation-api/src/types/project-info.ts
@@ -1,0 +1,5 @@
+export interface ProjectInfo {
+    projectRoot: string;
+    appName: string;
+    apps: string[];
+}