+++ packages/fiori-generator-shared/src/index.ts
@@ -1,1 +1,2 @@
 export * from './cap';
+export { generateReadMe } from './read-me';