+++ packages/preview-middleware-client/src/adp/controllers/AddFragment.controller.ts
@@ -2,7 +2,6 @@
 import Button from 'sap/m/Button';
 import type Dialog from 'sap/m/Dialog';
 import type ComboBox from 'sap/m/ComboBox';
-import MessageToast from 'sap/m/MessageToast';
 
 /** sap.ui.core */
 import type UI5Element from 'sap/ui/core/Element';
@@ -33,10 +32,18 @@ interface CreateFragmentProps {
     targetAggregation: string;
 }
 
+const radix = 10;
+
+type AddFragmentModel = JSONModel & {
+    getProperty(sPath: '/newFragmentName'): string;
+    getProperty(sPath: '/selectedIndex'): number;
+    getProperty(sPath: '/selectedAggregation/value'): string;
+};
+
 /**
  * @namespace open.ux.preview.client.adp.controllers
  */
-export default class AddFragment extends BaseDialog {
+export default class AddFragment extends BaseDialog<AddFragmentModel> {
     constructor(name: string, overlays: UI5Element, rta: RuntimeAuthoring) {
         super(name);
         this.rta = rta;
@@ -114,7 +121,7 @@ export default class AddFragment extends BaseDialog {
         );
 
         newSelectedControlChildren = newSelectedControlChildren.map((key) => {
-            return parseInt(key);
+            return parseInt(key, radix);
         });
 
         this.specialIndexHandling(selectedItemText);
@@ -135,7 +142,7 @@ export default class AddFragment extends BaseDialog {
         source.setEnabled(false);
 
         const fragmentName = this.model.getProperty('/newFragmentName');
-        const index = parseInt(this.model.getProperty('/selectedIndex'));
+        const index = this.model.getProperty('/selectedIndex');
         const targetAggregation = this.model.getProperty('/selectedAggregation/value');
         const fragmentData = {
             index,
@@ -182,7 +189,7 @@ export default class AddFragment extends BaseDialog {
         );
 
         selectedControlChildren = selectedControlChildren.map((key) => {
-            return parseInt(key);
+            return parseInt(key, radix);
         });
 
         this.model.setProperty('/selectedControlName', selectedControlName);
@@ -216,8 +223,7 @@ export default class AddFragment extends BaseDialog {
 
             this.model.setProperty('/fragmentList', fragments);
         } catch (e) {
-            MessageToast.show(e.message);
-            throw new Error(e.message);
+            this.handleError(e);
         }
 
         this.model.setProperty('/selectedIndex', indexArray.length - 1);