+++ packages/ui5-application-inquirer/src/types.ts
@@ -1,3 +1,4 @@
+import type { CdsUi5PluginInfo } from '@sap-ux/cap-config-writer';
 import type { CommonPromptOptions, PromptDefaultValue, UI5VersionChoice, YUIQuestion } from '@sap-ux/inquirer-common';
 import type { AutocompleteQuestionOptions } from 'inquirer-autocomplete-prompt';
 
@@ -81,6 +82,18 @@ type UI5VersionPromptOptions = {
 };
 
 /**
+ * Options for the enable TypeScript prompt. This allows for a default value to be determined based on the answers provided
+ * and additonal runtime cds information if available. This effectively constains the prompt options for the enable TypeScript prompt
+ * to be a function that returns a boolean value since enable TypeScript prompt default is conditional.
+ */
+type EnableTypeScriptPromptOptions = Omit<PromptDefaultValue<boolean>, 'default'> & {
+    /**
+     * Callback function to determine the default value for TypeScript
+     */
+    default?: (answers: UI5ApplicationAnswers & { capCdsInfo?: CdsUi5PluginInfo }) => boolean;
+};
+
+/**
  * These are boolean value prompt option keys
  */
 type booleanPromptKeys =
@@ -108,11 +121,7 @@ type DefaultValueInputPrompts =
     | promptNames.namespace
     | promptNames.ui5Version
     | promptNames.targetFolder;
-type DefaultValueConfirmPrompts =
-    | promptNames.enableCodeAssist
-    | promptNames.enableEslint
-    | promptNames.skipAnnotations
-    | promptNames.enableTypeScript;
+type DefaultValueConfirmPrompts = promptNames.enableCodeAssist | promptNames.enableEslint | promptNames.skipAnnotations;
 
 /**
  * Defines prompt/question default values and/or whether or not they should be shown.
@@ -147,7 +156,8 @@ type booleanValuePromptOtions = Record<
         validatorCallback?: (answer: boolean, promptName: string) => void;
     } & UI5ApplicationCommonPromptOptions
 > &
-    Record<DefaultValueConfirmPrompts, PromptDefaultValue<boolean>>;
+    Record<DefaultValueConfirmPrompts, PromptDefaultValue<boolean>> &
+    Record<promptNames.enableTypeScript, EnableTypeScriptPromptOptions>;
 
 export type UI5ApplicationQuestion = YUIQuestion<UI5ApplicationAnswers> &
     Partial<Pick<AutocompleteQuestionOptions, 'source'>>;