+++ packages/axios-extension/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @sap-ux/axios-extension
 
+## 1.14.4
+
+### Patch Changes
+
+-   4492fe10: fix for ui service generation response parsing
+
+## 1.14.3
+
+### Patch Changes
+
+-   d5d3626c: chore - Update to "qs": "6.11.0"
+
+## 1.14.2
+
+### Patch Changes
+
+-   65bfb244: Add Adaptation Project's Change Data Source generator prompting
+
+## 1.14.1
+
+### Patch Changes
+
+-   844e79c4: fix for v2 catalog services
+
+## 1.14.0
+
+### Minor Changes
+
+-   31cc53f8: Use new api endpoint for ui service generator
+
+## 1.13.1
+
+### Patch Changes
+
+-   869c1c0d: Prevents overwriting axios config params
+
+## 1.13.0
+
+### Minor Changes
+
+-   b2ee99fc: Updates how service specific annotations are requested
+
 ## 1.12.6
 
 ### Patch Changes