+++ packages/inquirer-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/inquirer-common
 
+## 0.3.1
+
+### Patch Changes
+
+-   65bfb244: Add Adaptation Project's Change Data Source generator prompting
+
+## 0.3.0
+
+### Minor Changes
+
+-   1ea831d5: Adds support for service url prompting
+
 ## 0.2.8
 
 ### Patch Changes